requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
confluent-kafka==2.3.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.0
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1
dbt-snowflake==1.7.0
apache-airflow==2.8.0
apache-airflow-providers-snowflake==5.1.0
fastapi==0.109.0
uvicorn==0.27.0
joblib==1.3.2
python-dotenv==1.0.0
boto3==1.34.0
avro-python3==1.10.2
requests==2.31.0
